sqlmodel-framework/examples/simple_postgres_bronze/pyproject.toml
[project]
name = "simple-postgres-bronze-example"
version = "0.1.0"
description = "Simple example of Bronze layer extraction using SQLModel Framework"
requires-python = ">=3.9"
dependencies = [
    "sqlmodel-framework @ git+https://github.com/Troubladore/airflow-data-platform.git@main#subdirectory=sqlmodel-framework",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]